Stubbed out worker handler.
